requirements.txt
boto3==1.26.165
json5==0.9.10
psycopg==3.1.9
psycopg2-binary==2.9.6
hashlib